refactor(cmd): Move utils to a dedicated modules
This will allow to easily distinct between real commands and utils. And
eventually to enforce more easily the independence of commands' code.

refs: #521
