+++ ggshield/cmd/auth/logout.py
@@ -4,7 +4,7 @@ import click
 from requests.exceptions import ConnectionError
 
 import ggshield.verticals.hmsl.utils as hmsl_utils
-from ggshield.cmd.common_options import add_common_options
+from ggshield.cmd.utils.common_options import add_common_options
 from ggshield.core.client import create_client
 from ggshield.core.config import Config
 from ggshield.core.errors import AuthError, UnexpectedError
