+++ ggshield/cmd/config/__init__.py
@@ -2,7 +2,7 @@ from typing import Any
 
 import click
 
-from ggshield.cmd.common_options import add_common_options
+from ggshield.cmd.utils.common_options import add_common_options
 
 from .config_get import config_get_cmd
 from .config_list import config_list_cmd
