+++ ggshield/cmd/config/config_unset.py
@@ -2,7 +2,7 @@ from typing import Any, Optional
 
 import click
 
-from ggshield.cmd.common_options import add_common_options
+from ggshield.cmd.utils.common_options import add_common_options
 from ggshield.core.config import Config
 
 from .config_set import set_user_config_field
