+++ ggshield/cmd/honeytoken/__init__.py
@@ -2,8 +2,8 @@ from typing import Any
 
 import click
 
-from ggshield.cmd.common_options import add_common_options
 from ggshield.cmd.honeytoken.create import create_cmd
+from ggshield.cmd.utils.common_options import add_common_options
 
 
 @click.group(commands={"create": create_cmd})
