+++ ggshield/cmd/iac/scan/precommit.py
@@ -3,13 +3,13 @@ from typing import Any, Optional, Sequence
 
 import click
 
-from ggshield.cmd.common_options import all_option, directory_argument
 from ggshield.cmd.iac.scan.all import display_iac_scan_all_result, iac_scan_all
 from ggshield.cmd.iac.scan.diff import display_iac_scan_diff_result, iac_scan_diff
 from ggshield.cmd.iac.scan.iac_scan_common_options import (
     add_iac_scan_common_options,
     update_context,
 )
+from ggshield.cmd.utils.common_options import all_option, directory_argument
 from ggshield.core.config import Config
 from ggshield.core.errors import handle_exception
 from ggshield.core.text_utils import display_warning
