+++ ggshield/cmd/sca/scan/ci.py
@@ -5,7 +5,6 @@ from typing import Any, Optional, Sequence
 import click
 from click import UsageError
 
-from ggshield.cmd.common_options import all_option, directory_argument
 from ggshield.cmd.sca.scan.sca_scan_utils import (
     create_output_handler,
     display_sca_beta_warning,
@@ -16,6 +15,7 @@ from ggshield.cmd.sca.scan.scan_common_options import (
     add_sca_scan_common_options,
     update_context,
 )
+from ggshield.cmd.utils.common_options import all_option, directory_argument
 from ggshield.core.config import Config
 from ggshield.core.errors import handle_exception
 from ggshield.core.git_hooks.ci import get_current_and_previous_state_from_ci_env
