+++ ggshield/cmd/sca/scan/diff.py
@@ -3,11 +3,6 @@ from typing import Any, Optional, Sequence
 
 import click
 
-from ggshield.cmd.common_options import (
-    directory_argument,
-    reference_option,
-    staged_option,
-)
 from ggshield.cmd.sca.scan.sca_scan_utils import (
     create_output_handler,
     display_sca_beta_warning,
@@ -16,6 +11,11 @@ from ggshield.cmd.sca.scan.sca_scan_utils import (
 from ggshield.cmd.sca.scan.scan_common_options import (
     add_sca_scan_common_options,
     update_context,
+)
+from ggshield.cmd.utils.common_options import (
+    directory_argument,
+    reference_option,
+    staged_option,
 )
 from ggshield.verticals.sca.collection.collection import (
     SCAScanDiffVulnerabilityCollection,
