+++ ggshield/cmd/utils/debug_logs.py
